rename folder for menus to lowercase
