+++ src/components/navigation/navbar/lecturer/courses/CourseMenu.vue
