+++ src/components/navigation/navbar/lecturer/courses/CourseMenuBody.vue
