+++ src/components/navigation/navbar/student/courses/CourseMenu.vue
