+++ src/components/navigation/navbar/student/courses/CourseMenuBody.vue
