+++ src/components/navigation/navbar/student/courses/CourseMenuFooter.vue
